Moved files to dim2 directory
